Rename PowerUP.h to game-source-code/PowerUP.h
